[HTM-755] Refactor MapController integration tests
some auth test are left for later
